test(gpu): focus evidence lanes on behavioral coverage
