+++ pyproject.toml
@@ -103,7 +103,7 @@ markers = [
     "requires_systemd: requires a systemd user manager",
     "requires_toxiproxy: requires the pinned Toxiproxy server binary",
     "requires_torch: requires the optional PyTorch provider",
-    "requires_compute_sanitizer: requires NVIDIA Compute Sanitizer and a compatible GPU",
+    "requires_compute_sanitizer: requires NVIDIA Compute Sanitizer, nvcc, and a compatible GPU",
     "requires_ncu: requires NVIDIA Nsight Compute and its official Python report interface",
     "requires_triton: requires the optional Triton provider and a compatible GPU",
     "requires_nvbench: requires a configured NVBench benchmark executable and compatible GPU",
