+++ tests/adapters/test_nvbench.py
@@ -2,22 +2,19 @@
 
 from __future__ import annotations
 
-import hashlib
 import json
 import struct
 from pathlib import Path
 from typing import cast
 
 import pytest
-from pydantic import ValidationError
 
 from flameox.adapters.builtins import build_capture_invocation
 from flameox.adapters.nvbench import NvbenchExtractor
 from flameox.adapters.options import bind_adapter_options
-from flameox.application import ImportService, NvbenchImportService
-from flameox.application.imports import BundleMember, ImportBundleRequest
+from flameox.application import NvbenchImportService
 from flameox.catalog import Catalog
-from flameox.domain import ArtifactKind, DomainError, ErrorCode
+from flameox.domain import DomainError, ErrorCode
 from flameox.storage import GenerationManifest, RunStore, Workspace
 
 
@@ -102,47 +99,28 @@ def _nvbench_json(
     )
 
 
-def _import_bundle(
-    workspace: Workspace,
-    json_path: Path,
-    sidecar_path: Path,
-    *,
-    sidecar_filename: str = "out.json-bin/0.bin",
-) -> str:
-    result = ImportService(workspace)._import_provider_bundle(
-        ImportBundleRequest(
-            primary=BundleMember(
-                path=json_path,
-                role="primary",
-                media_type="application/json",
-            ),
-            sidecars=(
-                BundleMember(
-                    path=sidecar_path,
-                    role="nvbench_sidecar",
-                    media_type="application/octet-stream",
-                    display_name=sidecar_filename,
-                ),
-            ),
-            kind=ArtifactKind.BENCHMARK_SAMPLES,
-            producer="nvbench",
-            producer_version="0.1.0",
-        )
-    )
-    return result.run.run_id
-
-
 def test_nvbench_extracts_float32_sample_times_without_lossy_conversion(
     tmp_path: Path,
 ) -> None:
     workspace = Workspace.initialize(tmp_path)
     samples = [0.004571, 0.004580, 0.004562]
     json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
-    json_path.write_text(_nvbench_json(sample_count=len(samples)))
+    sidecar_path = tmp_path / "out.json-bin" / "0.bin"
+    sidecar_path.parent.mkdir()
+    document = json.loads(_nvbench_json(sample_count=len(samples)))
+    document["benchmarks"][0]["states"].append(
+        {
+            "name": "[T=F32 Elements=2^20]",
+            "device": 0,
+            "is_skipped": True,
+            "skip_reason": "No CUDA device",
+            "summaries": [],
+        }
+    )
+    json_path.write_text(json.dumps(document))
     sidecar_path.write_bytes(_float32_bytes(samples))
 
-    run_id = _import_bundle(workspace, json_path, sidecar_path)
+    run_id = NvbenchImportService(workspace).import_json(json_path).run.run_id
 
     result = NvbenchExtractor(workspace).extract(run_id)
     repeated = NvbenchExtractor(workspace).extract(run_id)
@@ -191,10 +169,11 @@ def test_nvbench_extracts_float32_sample_times_without_lossy_conversion(
 def test_nvbench_rejects_nonfinite_binary_samples(tmp_path: Path) -> None:
     workspace = Workspace.initialize(tmp_path)
     json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
+    sidecar_path = tmp_path / "out.json-bin" / "0.bin"
+    sidecar_path.parent.mkdir()
     json_path.write_text(_nvbench_json(sample_count=1))
     sidecar_path.write_bytes(_float32_bytes([float("nan")]))
-    run_id = _import_bundle(workspace, json_path, sidecar_path)
+    run_id = NvbenchImportService(workspace).import_json(json_path).run.run_id
 
     with pytest.raises(DomainError, match="non-finite"):
         NvbenchExtractor(workspace).extract(run_id)
@@ -224,141 +203,20 @@ def test_nvbench_measurement_identity_includes_consumed_sidecar(tmp_path: Path) 
     assert len({measurement_id for _, measurement_id in measurement_ids}) == 2
 
 
-def test_nvbench_extracts_real_upstream_shape_with_nested_relative_paths(
+@pytest.mark.parametrize("invalid_size", ["-5", 3], ids=["negative", "numeric-json"])
+def test_nvbench_rejects_sidecar_sizes_outside_documented_encoding(
     tmp_path: Path,
+    invalid_size: object,
 ) -> None:
-    """Reflect real NVBench output: string-serialized size and nested
-    relative sidecar paths like ``out.json-bin/0.bin``.
-
-    NVBench's json_printer.cu writes int64 summary values as decimal
-    strings and stores sidecar filenames as paths relative to the JSON
-    file's directory (e.g. ``out.json-bin/0.bin``).  The bundle import
-    must preserve that relative path in ``display_name`` so the extractor
-    can match it exactly.
-    """
-    workspace = Workspace.initialize(tmp_path)
-    samples = [0.003120, 0.003145, 0.003108, 0.003129]
-    json_path = tmp_path / "out.json"
-    # Real NVBench writes sidecars into a sibling directory:
-    #   out.json-bin/0.bin
-    sidecar_dir = tmp_path / "out.json-bin"
-    sidecar_dir.mkdir()
-    sidecar_path = sidecar_dir / "0.bin"
-    json_path.write_text(
-        _nvbench_json(
-            sidecar_filename="out.json-bin/0.bin",
-            sample_count=len(samples),
-        )
-    )
-    sidecar_path.write_bytes(_float32_bytes(samples))
-
-    run_id = _import_bundle(
-        workspace,
-        json_path,
-        sidecar_path,
-        sidecar_filename="out.json-bin/0.bin",
-    )
-
-    result = NvbenchExtractor(workspace).extract(run_id)
-    assert result.measurement_count == 4
-    assert result.benchmark_count == 1
-    with Catalog(workspace).open_snapshot() as snapshot:
-        rows = snapshot.execute(
-            "SELECT value_float, unit FROM measurements WHERE run_id = ? ORDER BY value_index",
-            (run_id,),
-        ).fetchall()
-    assert len(rows) == 4
-    for row in rows:
-        value_float, unit = row
-        assert unit == "seconds"
-        assert value_float is not None
-    assert rows[0][0] == pytest.approx(0.003120, abs=1e-6)
-
-
-def test_nvbench_rejects_non_decimal_string_size(tmp_path: Path) -> None:
-    """NVBench serializes size as a decimal string; non-decimal values
-    must be rejected."""
     workspace = Workspace.initialize(tmp_path)
     json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
-    doc = json.loads(_nvbench_json(sample_count=3))
-    doc["benchmarks"][0]["states"][0]["summaries"][0]["data"][1]["value"] = "-5"
-    json_path.write_text(json.dumps(doc))
-    sidecar_path.write_bytes(_float32_bytes([0.01, 0.02, 0.03]))
-
-    run_id = _import_bundle(workspace, json_path, sidecar_path)
+    document = json.loads(_nvbench_json(sample_count=3))
+    document["benchmarks"][0]["states"][0]["summaries"][0]["data"][1]["value"] = invalid_size
+    json_path.write_text(json.dumps(document))
 
     with pytest.raises(DomainError) as error:
-        NvbenchExtractor(workspace).extract(run_id)
-    assert error.value.code is ErrorCode.ARTIFACT_PARSE_FAILED
+        NvbenchImportService(workspace).import_json(json_path, allow_external_path=True)
 
-
-def test_nvbench_rejects_numeric_sidecar_size_in_current_schema(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    doc = json.loads(_nvbench_json(sample_count=3))
-    doc["benchmarks"][0]["states"][0]["summaries"][0]["data"][1]["value"] = 3
-    json_path.write_text(json.dumps(doc))
-
-    with pytest.raises(DomainError) as error:
-        NvbenchImportService(workspace).import_json(json_path)
-
-    assert error.value.code is ErrorCode.ARTIFACT_PARSE_FAILED
-
-
-def test_nvbench_handles_skipped_states(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
-    doc = json.loads(_nvbench_json(sample_count=2))
-    doc["benchmarks"][0]["states"].append(
-        {
-            "name": "[T=F32 Elements=2^20]",
-            "device": 0,
-            "is_skipped": True,
-            "skip_reason": "No CUDA device",
-            "summaries": [],
-        }
-    )
-    json_path.write_text(json.dumps(doc))
-    sidecar_path.write_bytes(_float32_bytes([0.01, 0.02]))
-
-    run_id = _import_bundle(workspace, json_path, sidecar_path)
-
-    result = NvbenchExtractor(workspace).extract(run_id)
-    assert result.measurement_count == 2
-
-
-def test_nvbench_rejects_missing_sidecar(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    other_path = tmp_path / "wrong.bin"
-    json_path.write_text(_nvbench_json(sidecar_filename="0.bin", sample_count=2))
-    other_path.write_bytes(_float32_bytes([0.01, 0.02]))
-
-    run_id = _import_bundle(
-        workspace,
-        json_path,
-        other_path,
-        sidecar_filename="wrong.bin",
-    )
-
-    with pytest.raises(DomainError) as error:
-        NvbenchExtractor(workspace).extract(run_id)
-    assert error.value.code is ErrorCode.ARTIFACT_PARSE_FAILED
-
-
-def test_nvbench_rejects_sidecar_size_mismatch(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
-    json_path.write_text(_nvbench_json(sample_count=3))
-    sidecar_path.write_bytes(_float32_bytes([0.01, 0.02]))
-
-    run_id = _import_bundle(workspace, json_path, sidecar_path)
-
-    with pytest.raises(DomainError) as error:
-        NvbenchExtractor(workspace).extract(run_id)
     assert error.value.code is ErrorCode.ARTIFACT_PARSE_FAILED
 
 
@@ -367,7 +225,8 @@ def test_nvbench_rejects_declared_samples_over_row_budget_before_decoding(
 ) -> None:
     workspace = Workspace.initialize(tmp_path)
     json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
+    sidecar_path = tmp_path / "out.json-bin" / "0.bin"
+    sidecar_path.parent.mkdir()
     document = json.loads(_nvbench_json(sample_count=2))
     summaries = document["benchmarks"][0]["states"][0]["summaries"]
     summaries.append(
@@ -376,11 +235,13 @@ def test_nvbench_rejects_declared_samples_over_row_budget_before_decoding(
     summaries[1]["data"] = [dict(datum) for datum in summaries[1]["data"]]
     summaries[1]["data"][0]["value"] = "out.json-bin/1.bin"
     json_path.write_text(json.dumps(document))
-    # This payload is deliberately malformed for the declared sample count. A
-    # decoder-first implementation reports a size mismatch instead of rejecting
-    # the declared allocation against the generation budget.
-    sidecar_path.write_bytes(b"not-float32-data")
-    run_id = _import_bundle(workspace, json_path, sidecar_path)
+    # Import verifies the byte length, while extraction must reject the declared
+    # row allocation before decoding these non-finite values.
+    sidecar_path.write_bytes(_float32_bytes([float("nan"), float("nan")]))
+    second_sidecar = tmp_path / "out.json-bin" / "1.bin"
+    second_sidecar.parent.mkdir(parents=True, exist_ok=True)
+    second_sidecar.write_bytes(_float32_bytes([float("nan"), float("nan")]))
+    run_id = NvbenchImportService(workspace).import_json(json_path).run.run_id
     config = workspace.config.model_copy(
         update={
             "storage": workspace.config.storage.model_copy(update={"max_rows_per_generation": 3})
@@ -395,226 +256,15 @@ def test_nvbench_rejects_declared_samples_over_row_budget_before_decoding(
     assert error.value.details == {"rows": 4, "max_rows": 3}
 
 
-def test_nvbench_rejects_malformed_json(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
-    json_path.write_text("{not valid json")
-    sidecar_path.write_bytes(_float32_bytes([0.01]))
-
-    run_id = _import_bundle(workspace, json_path, sidecar_path)
-
-    with pytest.raises(DomainError) as error:
-        NvbenchExtractor(workspace).extract(run_id)
-    assert error.value.code is ErrorCode.ARTIFACT_PARSE_FAILED
-
-
 def test_nvbench_rejects_empty_document_without_schema_identity(
     tmp_path: Path,
 ) -> None:
     workspace = Workspace.initialize(tmp_path)
     json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
     json_path.write_text('{"meta": {}, "benchmarks": []}')
-    sidecar_path.write_bytes(_float32_bytes([0.01]))
-
-    run_id = _import_bundle(workspace, json_path, sidecar_path)
 
     with pytest.raises(DomainError, match="schema major missing"):
-        NvbenchExtractor(workspace).extract(run_id)
-
-
-def test_nvbench_rejects_run_without_primary(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
-    json_path.write_text(_nvbench_json(sample_count=1))
-    sidecar_path.write_bytes(_float32_bytes([0.01]))
-
-    result = ImportService(workspace)._import_provider_bundle(
-        ImportBundleRequest(
-            primary=BundleMember(
-                path=json_path,
-                role="secondary",
-                media_type="application/json",
-            ),
-            sidecars=(
-                BundleMember(
-                    path=sidecar_path,
-                    role="nvbench_sidecar",
-                    media_type="application/octet-stream",
-                ),
-            ),
-            kind=ArtifactKind.BENCHMARK_SAMPLES,
-            producer="nvbench",
-        )
-    )
-
-    with pytest.raises(DomainError) as error:
-        NvbenchExtractor(workspace).extract(result.run.run_id)
-    assert error.value.code is ErrorCode.ARTIFACT_PARSE_FAILED
-
-
-def test_bundle_import_rejects_excess_members(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    json_path.write_text(_nvbench_json(sample_count=1))
-    sidecar_files: list[BundleMember] = []
-    for i in range(100):
-        sidecar = tmp_path / f"{i}.bin"
-        sidecar.write_bytes(_float32_bytes([float(i)]))
-        sidecar_files.append(BundleMember(path=sidecar, role="nvbench_sidecar"))
-
-    with pytest.raises(ValidationError):
-        ImportService(workspace)._import_provider_bundle(
-            ImportBundleRequest(
-                primary=BundleMember(path=json_path),
-                sidecars=tuple(sidecar_files),
-                kind=ArtifactKind.BENCHMARK_SAMPLES,
-                producer="nvbench",
-            )
-        )
-
-
-def test_bundle_import_rejects_duplicate_sidecar(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
-    json_path.write_text(_nvbench_json(sample_count=1))
-    sidecar_path.write_bytes(_float32_bytes([0.01]))
-
-    with pytest.raises(DomainError) as error:
-        ImportService(workspace)._import_provider_bundle(
-            ImportBundleRequest(
-                primary=BundleMember(path=json_path),
-                sidecars=(
-                    BundleMember(path=sidecar_path, role="nvbench_sidecar"),
-                    BundleMember(path=sidecar_path, role="nvbench_sidecar"),
-                ),
-                kind=ArtifactKind.BENCHMARK_SAMPLES,
-                producer="nvbench",
-            )
-        )
-    assert error.value.code is ErrorCode.EXECUTION_REFUSED
-
-
-def test_bundle_import_allows_distinct_members_with_identical_content(tmp_path: Path) -> None:
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    sidecar_a = tmp_path / "a.bin"
-    sidecar_b = tmp_path / "b.bin"
-    json_path.write_text(_nvbench_json(sample_count=1))
-    payload = _float32_bytes([0.01])
-    sidecar_a.write_bytes(payload)
-    sidecar_b.write_bytes(payload)
-
-    result = ImportService(workspace)._import_provider_bundle(
-        ImportBundleRequest(
-            primary=BundleMember(path=json_path),
-            sidecars=(
-                BundleMember(path=sidecar_a, role="nvbench_sidecar"),
-                BundleMember(path=sidecar_b, role="nvbench_sidecar"),
-            ),
-            kind=ArtifactKind.BENCHMARK_SAMPLES,
-            producer="nvbench",
-        )
-    )
-
-    assert result.sidecar_artifact_ids[0] == result.sidecar_artifact_ids[1]
-
-
-def test_bundle_import_verifies_declared_byte_length(tmp_path: Path) -> None:
-    """A source that changes size between manifest and import is rejected."""
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
-    json_path.write_text(_nvbench_json(sample_count=1))
-    payload = _float32_bytes([0.01])
-    sidecar_path.write_bytes(payload)
-
-    # Declare a wrong byte length to simulate a raced source
-    with pytest.raises(DomainError) as error:
-        ImportService(workspace)._import_provider_bundle(
-            ImportBundleRequest(
-                primary=BundleMember(path=json_path),
-                sidecars=(
-                    BundleMember(
-                        path=sidecar_path,
-                        role="nvbench_sidecar",
-                        display_name="out.json-bin/0.bin",
-                        expected_byte_length=len(payload) + 1,
-                    ),
-                ),
-                kind=ArtifactKind.BENCHMARK_SAMPLES,
-                producer="nvbench",
-            )
-        )
-    assert error.value.code is ErrorCode.ARTIFACT_INTEGRITY_FAILED
-    assert "byte length mismatch" in error.value.message
-
-
-def test_bundle_import_verifies_declared_sha256(tmp_path: Path) -> None:
-    """A source whose digest differs from the manifest is rejected."""
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
-    json_path.write_text(_nvbench_json(sample_count=1))
-    payload = _float32_bytes([0.01])
-    sidecar_path.write_bytes(payload)
-
-    wrong_sha = "sha256:" + "0" * 64
-    with pytest.raises(DomainError) as error:
-        ImportService(workspace)._import_provider_bundle(
-            ImportBundleRequest(
-                primary=BundleMember(path=json_path),
-                sidecars=(
-                    BundleMember(
-                        path=sidecar_path,
-                        role="nvbench_sidecar",
-                        display_name="out.json-bin/0.bin",
-                        expected_sha256=wrong_sha,
-                    ),
-                ),
-                kind=ArtifactKind.BENCHMARK_SAMPLES,
-                producer="nvbench",
-            )
-        )
-    assert error.value.code is ErrorCode.ARTIFACT_INTEGRITY_FAILED
-    assert "sha256 mismatch" in error.value.message
-
-
-def test_bundle_import_accepts_correct_declared_integrity(tmp_path: Path) -> None:
-    """Correct declared byte length and sha256 pass verification."""
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "nvbench.json"
-    sidecar_path = tmp_path / "0.bin"
-    json_path.write_text(_nvbench_json(sample_count=1))
-    payload = _float32_bytes([0.01])
-    sidecar_path.write_bytes(payload)
-    actual_sha = "sha256:" + hashlib.sha256(payload).hexdigest()
-
-    result = ImportService(workspace)._import_provider_bundle(
-        ImportBundleRequest(
-            primary=BundleMember(path=json_path),
-            sidecars=(
-                BundleMember(
-                    path=sidecar_path,
-                    role="nvbench_sidecar",
-                    display_name="out.json-bin/0.bin",
-                    expected_byte_length=len(payload),
-                    expected_sha256=actual_sha,
-                ),
-            ),
-            kind=ArtifactKind.BENCHMARK_SAMPLES,
-            producer="nvbench",
-        )
-    )
-    assert result.primary_artifact_id
-
-
-# ---------------------------------------------------------------------------
-# NvbenchImportService: provider-defined selection boundary
-# ---------------------------------------------------------------------------
+        NvbenchImportService(workspace).import_json(json_path, allow_external_path=True)
 
 
 def test_nvbench_import_service_imports_only_declared_sidecars(tmp_path: Path) -> None:
@@ -647,6 +297,34 @@ def test_nvbench_import_service_imports_only_declared_sidecars(tmp_path: Path) -
     assert "out.json-bin/1.bin" not in display_names
 
 
+def test_nvbench_import_rejects_bundles_over_public_member_limit(tmp_path: Path) -> None:
+    workspace = Workspace.initialize(tmp_path)
+    json_path = tmp_path / "out.json"
+    document = json.loads(_nvbench_json(sample_count=1))
+    summary = document["benchmarks"][0]["states"][0]["summaries"][0]
+    document["benchmarks"][0]["states"][0]["summaries"] = [
+        {
+            **summary,
+            "data": [
+                {
+                    "name": "filename",
+                    "type": "string",
+                    "value": f"out.json-bin/{index}.bin",
+                },
+                {"name": "size", "type": "int64", "value": "1"},
+            ],
+        }
+        for index in range(100)
+    ]
+    json_path.write_text(json.dumps(document))
+
+    with pytest.raises(DomainError) as error:
+        NvbenchImportService(workspace).import_json(json_path, allow_external_path=True)
+
+    assert error.value.code is ErrorCode.EXECUTION_REFUSED
+    assert "exceeding the 99-sidecar limit" in error.value.message
+
+
 def test_nvbench_import_service_binds_expected_byte_length(tmp_path: Path) -> None:
     """Sidecar expected_byte_length = count * 4 is verified after import."""
     workspace = Workspace.initialize(tmp_path)
@@ -664,33 +342,15 @@ def test_nvbench_import_service_binds_expected_byte_length(tmp_path: Path) -> No
     assert "byte length mismatch" in error.value.message
 
 
-def test_nvbench_import_service_round_trip_extract(tmp_path: Path) -> None:
-    """import_json followed by NvbenchExtractor.extract produces correct measurements."""
-    workspace = Workspace.initialize(tmp_path)
-    samples = [0.003120, 0.003145, 0.003108, 0.003129]
-    json_path = tmp_path / "out.json"
-    sidecar_dir = tmp_path / "out.json-bin"
-    sidecar_dir.mkdir()
-    sidecar_path = sidecar_dir / "0.bin"
-    json_path.write_text(_nvbench_json(sample_count=len(samples)))
-    sidecar_path.write_bytes(_float32_bytes(samples))
-
-    result = NvbenchImportService(workspace).import_json(json_path, allow_external_path=True)
-    extracted = NvbenchExtractor(workspace).extract(result.run.run_id)
-    assert extracted.measurement_count == 4
-    assert extracted.benchmark_count == 1
-
-
 def test_nvbench_rejects_unverified_json_schema_major_before_decoding(tmp_path: Path) -> None:
     workspace = Workspace.initialize(tmp_path)
     json_path = tmp_path / "out.json"
     sidecar_path = tmp_path / "0.bin"
     json_path.write_text(_nvbench_json(sidecar_filename="0.bin", json_version=(2, 0, 0)))
     sidecar_path.write_bytes(_float32_bytes([0.01, 0.02, 0.03]))
-    run_id = _import_bundle(workspace, json_path, sidecar_path, sidecar_filename="0.bin")
 
     with pytest.raises(DomainError, match="only major 1 is verified"):
-        NvbenchExtractor(workspace).extract(run_id)
+        NvbenchImportService(workspace).import_json(json_path, allow_external_path=True)
 
 
 def test_nvbench_import_service_rejects_malformed_json(tmp_path: Path) -> None:
@@ -825,60 +485,43 @@ def test_nvbench_import_service_rejects_unknown_file_hint(
     assert "Unknown NVBench file hint" in error.value.message
 
 
-def test_nvbench_import_service_rejects_known_hint_missing_filename(
+@pytest.mark.parametrize(
+    ("missing_name", "remaining_datum"),
+    [
+        ("filename", {"name": "size", "type": "int64", "value": "3"}),
+        (
+            "size",
+            {
+                "name": "filename",
+                "type": "string",
+                "value": "out.json-bin/0.bin",
+            },
+        ),
+    ],
+)
+def test_nvbench_import_rejects_incomplete_sidecar_declarations(
     tmp_path: Path,
+    missing_name: str,
+    remaining_datum: dict[str, str],
 ) -> None:
-    """A known file hint missing the filename datum must be rejected."""
     workspace = Workspace.initialize(tmp_path)
     json_path = tmp_path / "out.json"
-    doc = json.loads(_nvbench_json(sample_count=3))
-    state = doc["benchmarks"][0]["states"][0]
-    state["summaries"].append(
+    document = json.loads(_nvbench_json(sample_count=3))
+    document["benchmarks"][0]["states"][0]["summaries"].append(
         {
-            "tag": "sample_freqs_no_filename",
-            "hint": "file/sample_freqs",
-            "data": [
-                {"name": "size", "type": "int64", "value": "3"},
-            ],
-        }
-    )
-    json_path.write_text(json.dumps(doc))
-
-    with pytest.raises(DomainError) as error:
-        NvbenchImportService(workspace).import_json(json_path, allow_external_path=True)
-    assert error.value.code is ErrorCode.ARTIFACT_PARSE_FAILED
-    assert "missing" in error.value.message
-    assert "filename" in error.value.message
-
-
-def test_nvbench_import_service_rejects_known_hint_missing_size(
-    tmp_path: Path,
-) -> None:
-    """A known file hint missing the size datum must be rejected."""
-    workspace = Workspace.initialize(tmp_path)
-    json_path = tmp_path / "out.json"
-    sidecar_dir = tmp_path / "out.json-bin"
-    sidecar_dir.mkdir()
-    sidecar_path = sidecar_dir / "0.bin"
-    doc = json.loads(_nvbench_json(sample_count=3))
-    state = doc["benchmarks"][0]["states"][0]
-    state["summaries"].append(
-        {
-            "tag": "sample_times_no_size",
+            "tag": f"sample_times_no_{missing_name}",
             "hint": "file/sample_times",
-            "data": [
-                {"name": "filename", "type": "string", "value": "out.json-bin/0.bin"},
-            ],
+            "data": [remaining_datum],
         }
     )
-    json_path.write_text(json.dumps(doc))
-    sidecar_path.write_bytes(_float32_bytes([0.01, 0.02, 0.03]))
+    json_path.write_text(json.dumps(document))
 
     with pytest.raises(DomainError) as error:
         NvbenchImportService(workspace).import_json(json_path, allow_external_path=True)
+
     assert error.value.code is ErrorCode.ARTIFACT_PARSE_FAILED
     assert "missing" in error.value.message
-    assert "size" in error.value.message
+    assert missing_name in error.value.message
 
 
 def test_nvbench_import_service_rejects_missing_sidecar(tmp_path: Path) -> None:
@@ -890,12 +533,9 @@ def test_nvbench_import_service_rejects_missing_sidecar(tmp_path: Path) -> None:
 
     with pytest.raises(DomainError) as error:
         NvbenchImportService(workspace).import_json(json_path, allow_external_path=True)
-    # The missing sidecar triggers either EXECUTION_REFUSED (preflight)
-    # or ARTIFACT_INTEGRITY_FAILED (import), both are acceptable
-    assert error.value.code in {
-        ErrorCode.EXECUTION_REFUSED,
-        ErrorCode.ARTIFACT_INTEGRITY_FAILED,
-    }
+
+    assert error.value.code is ErrorCode.EXECUTION_REFUSED
+    assert "out.json-bin/0.bin" in error.value.message
 
 
 def test_nvbench_import_service_rejects_sidecar_traversal(tmp_path: Path) -> None:
@@ -941,12 +581,12 @@ def test_nvbench_import_rejects_symlinked_primary_before_parsing(tmp_path: Path)
 
 def test_nvbench_import_applies_provider_document_limit(
     tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     workspace = Workspace.initialize(tmp_path)
     json_path = tmp_path / "out.json"
     json_path.write_text(_nvbench_json())
-    monkeypatch.setattr("flameox.application.nvbench_imports._MAX_NVBENCH_DOCUMENT_BYTES", 8)
+    with json_path.open("ab") as document:
+        document.truncate(16 * 1024 * 1024 + 1)
 
     with pytest.raises(DomainError) as error:
         NvbenchImportService(workspace).import_json(json_path)
@@ -973,27 +613,6 @@ def test_nvbench_import_service_verifies_primary_sha256(tmp_path: Path) -> None:
     assert "sha256 mismatch" in error.value.message
 
 
-def test_nvbench_import_service_accepts_correct_primary_sha256(
-    tmp_path: Path,
-) -> None:
-    """A correct expected_sha256 on the primary JSON passes verification."""
-    workspace = Workspace.initialize(tmp_path)
-    samples = [0.01, 0.02]
-    json_path = tmp_path / "out.json"
-    sidecar_dir = tmp_path / "out.json-bin"
-    sidecar_dir.mkdir()
-    sidecar_path = sidecar_dir / "0.bin"
-    json_content = _nvbench_json(sample_count=len(samples))
-    json_path.write_text(json_content)
-    sidecar_path.write_bytes(_float32_bytes(samples))
-    correct_sha = "sha256:" + hashlib.sha256(json_content.encode()).hexdigest()
-
-    result = NvbenchImportService(workspace).import_json(
-        json_path, allow_external_path=True, expected_sha256=correct_sha
-    )
-    assert result.primary_artifact_id
-
-
 def test_nvbench_capture_invocation_uses_jsonbin_by_default(tmp_path: Path) -> None:
     bound = bind_adapter_options("nvbench", None, project_root=tmp_path)
     invocation = build_capture_invocation(
@@ -1003,16 +622,13 @@ def test_nvbench_capture_invocation_uses_jsonbin_by_default(tmp_path: Path) -> N
         executable=None,
         options=cast(dict[str, object], bound),
     )
-    argv = invocation.argv
-    # argv[0] is the benchmark executable, not a wrapper
-    assert argv[0] == "./bench"
-    # --jsonbin is the default; --json must NOT be present (alternative modes)
-    assert "--jsonbin" in argv
-    assert "--json" not in argv
-    # No "--" separator — flags are injected into the benchmark argv
-    assert "--" not in argv
-    # Workload's own args appear after injected flags
-    assert argv[-2:] == ("-a", "T=I32")
+    assert invocation.argv == (
+        "./bench",
+        "--jsonbin",
+        str(tmp_path / "output" / "nvbench.json"),
+        "-a",
+        "T=I32",
+    )
 
 
 def test_nvbench_capture_invocation_falls_back_to_json_without_jsonbin(
@@ -1030,10 +646,11 @@ def test_nvbench_capture_invocation_falls_back_to_json_without_jsonbin(
         executable=None,
         options=cast(dict[str, object], bound),
     )
-    argv = invocation.argv
-    assert "--json" in argv
-    assert "--jsonbin" not in argv
-    assert argv[0] == "./bench"
+    assert invocation.argv == (
+        "./bench",
+        "--json",
+        str(tmp_path / "output" / "nvbench.json"),
+    )
 
 
 def test_nvbench_capture_invocation_passes_optional_flags(tmp_path: Path) -> None:
@@ -1055,73 +672,39 @@ def test_nvbench_capture_invocation_passes_optional_flags(tmp_path: Path) -> Non
         executable=None,
         options=cast(dict[str, object], bound),
     )
-    argv = invocation.argv
-    # Only --jsonbin, not --json
-    assert "--jsonbin" in argv
-    assert "--json" not in argv
-    assert "--stopping-criterion" in argv
-    assert "entropy" in argv
-    assert "--min-samples" in argv
-    assert "10" in argv
-    assert "--timeout" in argv
-    assert "30.0" in argv
-    assert "-d" in argv
-    assert "0" in argv
+    assert invocation.argv == (
+        "./bench",
+        "--jsonbin",
+        str(tmp_path / "output" / "nvbench.json"),
+        "--stopping-criterion",
+        "entropy",
+        "--min-samples",
+        "10",
+        "--timeout",
+        "30.0",
+        "-d",
+        "0",
+    )
 
 
-def test_nvbench_capture_invocation_rejects_conflicting_json_flag(
+@pytest.mark.parametrize(
+    "conflicting_arguments",
+    [
+        ("--json", "other.json"),
+        ("--jsonbin", "other.json"),
+        ("--json=other.json",),
+        ("--jsonbin=other.json",),
+    ],
+)
+def test_nvbench_capture_rejects_caller_owned_output(
     tmp_path: Path,
+    conflicting_arguments: tuple[str, ...],
 ) -> None:
     bound = bind_adapter_options("nvbench", None, project_root=tmp_path)
     with pytest.raises(DomainError) as error:
         build_capture_invocation(
             "nvbench",
-            ("./bench", "--json", "other.json"),
-            tmp_path / "output",
-            executable=None,
-            options=cast(dict[str, object], bound),
-        )
-    assert error.value.code is ErrorCode.INVALID_CAPTURE_PLAN
-
-
-def test_nvbench_capture_invocation_rejects_conflicting_jsonbin_flag(
-    tmp_path: Path,
-) -> None:
-    bound = bind_adapter_options("nvbench", None, project_root=tmp_path)
-    with pytest.raises(DomainError) as error:
-        build_capture_invocation(
-            "nvbench",
-            ("./bench", "--jsonbin", "other.json"),
-            tmp_path / "output",
-            executable=None,
-            options=cast(dict[str, object], bound),
-        )
-    assert error.value.code is ErrorCode.INVALID_CAPTURE_PLAN
-
-
-def test_nvbench_capture_invocation_rejects_equals_form_json(
-    tmp_path: Path,
-) -> None:
-    bound = bind_adapter_options("nvbench", None, project_root=tmp_path)
-    with pytest.raises(DomainError) as error:
-        build_capture_invocation(
-            "nvbench",
-            ("./bench", "--json=other.json"),
-            tmp_path / "output",
-            executable=None,
-            options=cast(dict[str, object], bound),
-        )
-    assert error.value.code is ErrorCode.INVALID_CAPTURE_PLAN
-
-
-def test_nvbench_capture_invocation_rejects_equals_form_jsonbin(
-    tmp_path: Path,
-) -> None:
-    bound = bind_adapter_options("nvbench", None, project_root=tmp_path)
-    with pytest.raises(DomainError) as error:
-        build_capture_invocation(
-            "nvbench",
-            ("./bench", "--jsonbin=other.json"),
+            ("./bench", *conflicting_arguments),
             tmp_path / "output",
             executable=None,
             options=cast(dict[str, object], bound),
