+++ tests/application/test_capture_nvbench_provider.py
@@ -7,12 +7,11 @@ from pathlib import Path
 
 import pytest
 
+from flameox.adapters.nvbench import NvbenchExtractor
 from flameox.application import CaptureResult, CaptureService, ExecutionPolicy
 from flameox.domain import (
     ArtifactKind,
     ArtifactRegistration,
-    CapabilityReport,
-    CapabilityStatus,
     CaptureStatus,
     ExecutionStatus,
 )
@@ -21,20 +20,6 @@ from tests.support.capture import disable_containment
 
 
 def _fake_nvbench_bench(path: Path) -> None:
-    """Write a fake NVBench benchmark executable.
-
-    The script reads ``--jsonbin <path>`` from argv, writes a JSON document
-    to ``<path>``, and writes ``sample_count`` float32 values to
-    ``<path>-bin/0.bin``.  An undeclared sibling ``<path>-bin/1.bin`` is
-    also written to verify it is NOT registered.
-
-    Modes:
-    - ``mode=ok``: exit 0, write valid JSON + sidecar + undeclared sibling.
-    - ``mode=malformed``: exit 0, write invalid JSON (no sidecars).
-    - ``mode=missing_sidecar``: exit 0, write valid JSON but no sidecar dir.
-    - ``mode=nonzero``: exit 7, write valid JSON + sidecar + undeclared sibling.
-    - ``mode=nonzero_malformed``: exit 7, write invalid JSON.
-    """
     path.write_text(
         """#!/usr/bin/env python3
 import json as _json
@@ -111,10 +96,7 @@ else:
         (sidecar_dir / "0.bin").write_bytes(
             struct.pack(f"<{sample_count}f", *samples)
         )
-        # Undeclared sibling that must NOT be registered
-        (sidecar_dir / "1.bin").write_bytes(
-            struct.pack("<2f", 0.99, 0.88)
-        )
+        (sidecar_dir / "1.bin").write_bytes(struct.pack("<2f", 0.99, 0.88))
 
 if mode.startswith("nonzero"):
     raise SystemExit(7)
@@ -123,35 +105,8 @@ if mode.startswith("nonzero"):
     path.chmod(path.stat().st_mode | stat.S_IXUSR)
 
 
-def _make_service(
-    workspace: Workspace,
-    bench: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> CaptureService:
-    report = CapabilityReport(
-        adapter="nvbench",
-        status=CapabilityStatus.AVAILABLE,
-        executable=str(bench),
-        version="fake-nvbench",
-        supported_modes=("benchmark",),
-        supported_formats=("nvbench-json", "nvbench-jsonbin"),
-        permission_status="granted",
-        probe_kind="active",
-    )
-    service = CaptureService(workspace)
-    monkeypatch.setattr(service.capabilities, "get", lambda _adapter: report)
-
-    async def probe(_adapter: str, *, refresh: bool = False) -> CapabilityReport:
-        assert refresh
-        return report
-
-    monkeypatch.setattr(service.capabilities, "probe", probe)
-    return service
-
-
 async def _capture(
     tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
     *,
     mode: str,
 ) -> tuple[CaptureService, CaptureResult]:
@@ -167,7 +122,7 @@ argv = ["./fake-bench", "mode={mode}"]
 timeout_seconds = 10
 """
     )
-    service = _make_service(workspace, bench, monkeypatch)
+    service = CaptureService(workspace)
     plan = await service.plan(
         workload_name="bench",
         adapter="nvbench",
@@ -183,9 +138,8 @@ def _benchmark_registrations(result: CaptureResult) -> list[ArtifactRegistration
 @pytest.mark.anyio
 async def test_nvbench_successful_capture_registers_declared_sidecars(
     tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    _, result = await _capture(tmp_path, monkeypatch, mode="ok")
+    service, result = await _capture(tmp_path, mode="ok")
 
     assert result.run.execution_status is ExecutionStatus.SUCCEEDED
     assert result.run.capture_status is CaptureStatus.REGISTERED
@@ -195,85 +149,47 @@ async def test_nvbench_successful_capture_registers_declared_sidecars(
     assert "nvbench_sidecar" in roles
     sidecar_regs = [item for item in benchmark_regs if item.role == "nvbench_sidecar"]
     assert len(sidecar_regs) == 1
-    # display_name must match the declared relative path
     assert sidecar_regs[0].display_name == "nvbench.json-bin/0.bin"
-    # The undeclared sibling (1.bin) must not be registered
     display_names = {item.display_name for item in benchmark_regs}
     assert "nvbench.json-bin/1.bin" not in display_names
+    extracted = NvbenchExtractor(service.workspace).extract(result.run.run_id)
+    assert extracted.measurement_count == 3
+    assert extracted.benchmark_count == 1
 
 
+@pytest.mark.parametrize(
+    ("mode", "expected_roles", "expected_limitation_codes"),
+    (
+        ("malformed", {"primary"}, {"expected_output_invalid"}),
+        ("missing_sidecar", {"primary"}, {"expected_output_invalid"}),
+        ("nonzero", {"primary", "nvbench_sidecar"}, {"nonzero_exit"}),
+        (
+            "nonzero_malformed",
+            {"primary"},
+            {"nonzero_exit", "expected_output_invalid"},
+        ),
+    ),
+    ids=(
+        "malformed-success",
+        "missing-sidecar-success",
+        "valid-nonzero",
+        "malformed-nonzero",
+    ),
+)
 @pytest.mark.anyio
-async def test_nvbench_successful_malformed_json_is_preserved_as_failed_attempt(
+async def test_nvbench_failed_attempts_preserve_only_valid_partial_artifacts(
     tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
+    mode: str,
+    expected_roles: set[str],
+    expected_limitation_codes: set[str],
 ) -> None:
-    service, result = await _capture(tmp_path, monkeypatch, mode="malformed")
+    service, result = await _capture(tmp_path, mode=mode)
 
     assert result.run.execution_status is ExecutionStatus.FAILED
     assert result.run.capture_status is CaptureStatus.REGISTERED
-    assert [item.role for item in _benchmark_registrations(result)] == ["primary"]
+    assert {item.role for item in _benchmark_registrations(result)} == expected_roles
     assert not list(service.workspace.paths.quarantine.glob("*/manifest.json"))
-    assert any(detail.code == "expected_output_invalid" for detail in result.run.limitation_details)
-
-
-@pytest.mark.anyio
-async def test_nvbench_successful_missing_sidecar_is_preserved_as_failed_attempt(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    service, result = await _capture(tmp_path, monkeypatch, mode="missing_sidecar")
-
-    assert result.run.execution_status is ExecutionStatus.FAILED
-    assert result.run.capture_status is CaptureStatus.REGISTERED
-    assert [item.role for item in _benchmark_registrations(result)] == ["primary"]
-    assert not list(service.workspace.paths.quarantine.glob("*/manifest.json"))
-
-
-@pytest.mark.anyio
-async def test_nvbench_nonzero_preserves_partial_sidecars(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    service, result = await _capture(tmp_path, monkeypatch, mode="nonzero")
-
-    assert result.run.execution_status is ExecutionStatus.FAILED
-    # Partial artifacts are preserved (not quarantined)
-    assert result.run.capture_status is CaptureStatus.REGISTERED
-    benchmark_regs = _benchmark_registrations(result)
-    roles = {item.role for item in benchmark_regs}
-    assert "primary" in roles
-    assert "nvbench_sidecar" in roles
-    sidecar_regs = [item for item in benchmark_regs if item.role == "nvbench_sidecar"]
-    assert len(sidecar_regs) == 1
-    assert sidecar_regs[0].display_name == "nvbench.json-bin/0.bin"
-    # Undeclared sibling must not be registered
-    display_names = {item.display_name for item in benchmark_regs}
-    assert "nvbench.json-bin/1.bin" not in display_names
-    # Must have nonzero limitation; valid output has no expected_output_invalid
     limitation_codes = {detail.code for detail in result.run.limitation_details}
-    assert "nonzero_exit" in limitation_codes
-    # Must NOT be quarantined
-    manifests = list(service.workspace.paths.quarantine.glob("*/manifest.json"))
-    assert not manifests
-
-
-@pytest.mark.anyio
-async def test_nvbench_nonzero_malformed_preserves_json_only(
-    tmp_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    service, result = await _capture(tmp_path, monkeypatch, mode="nonzero_malformed")
-
-    assert result.run.execution_status is ExecutionStatus.FAILED
-    assert result.run.capture_status is CaptureStatus.REGISTERED
-    benchmark_regs = _benchmark_registrations(result)
-    # The partial JSON is preserved as primary, but no sidecars
-    roles = {item.role for item in benchmark_regs}
-    assert "primary" in roles
-    assert "nvbench_sidecar" not in roles
-    # Must NOT be quarantined (partial preservation on nonzero)
-    manifests = list(service.workspace.paths.quarantine.glob("*/manifest.json"))
-    assert not manifests
-    limitation_codes = {detail.code for detail in result.run.limitation_details}
-    assert "nonzero_exit" in limitation_codes
-    assert "expected_output_invalid" in limitation_codes
+    assert limitation_codes & {"nonzero_exit", "expected_output_invalid"} == (
+        expected_limitation_codes
+    )
