+++ tests/fixtures/nsight_compute/corrupt.py.txt
@@ -1,0 +1,2 @@
+def load_report(path):
+    raise RuntimeError("corrupt official report")
